cyl-opt-local edited for testing checkpoints

+++ yales2/cyl-opt-local/pymooCFD/util/handleData.py
@@ -1,34 +1,53 @@
 import tarfile
-from pymooCFD.setupOpt import checkpointFile, dataDir, problem, nCP
+from pymooCFD.setupOpt import checkpointFile, dataDir, problem, nCP, archDir
 import shutil
 import numpy as np
+import time
+import os
 
 
-def archive(dirToComp, archDir, background=True):
+def archive(dirToComp, archDir=archDir, background=True):
+    try:
+        if len(os.listdir(dirToComp)) == 0:
+            # print('Directory given to archive is empty!!!')
+            # return
+            raise Exception('Directory given to archive is empty!!!')
+    except FileNotFoundError as err:
+        print(err)
+        # print('Directory given to archive does not exist!!!')
+        # return
+        raise Exception('Directory given to archive does not exist!!!')
     if background == True:
         from multiprocessing import Process
-        p = Process(target=compressDir, args=(dirToComp, ))
+        p = Process(target=compressDir, args=(dirToComp, archDir))
         p.start()
     else:
         compressDir(dirToComp, archDir)
-    
+
 
 def compressDir(dirToComp, archDir):
     print(f'{dirToComp} compression started')
+    # destination file naming
+    timestr = time.strftime("%y%m%d-%H%M%S")
     try:
-        fname = dirToComp[dirToComp.rindex("/"):]
+        fname = f'{dirToComp[dirToComp.rindex("/"):]}_{timestr}'
     except ValueError:
-        fname = dirToComp
+        fname = f'{dirToComp}_{timestr}'
+    # concatenate compression file path and name
     compFile = f'{archDir}/{fname}.tar.gz'
+    # write tar file using gzip compression algorithm
     with tarfile.open(compFile, 'w:gz') as tar:
         tar.add(dirToComp)
-    print(f'{dirToComp} compression finished')
+    print(f'{dirToComp} compressed into {compFile}')
     removeDir(dirToComp)
-    
+
 def removeDir(path):
+    if len(os.listdir(path)) == 0:
+        print('Directory given to remove is empty!!!')
+        return
     print(f'removing {path}..')
     try:
-        shutil.rmtree(path)        
+        shutil.rmtree(path)
         print(f"{path} removed successfully")
     except OSError as err:
         print(err)
@@ -59,10 +78,11 @@ def loadCP(checkpointFile=checkpointFile, hasTerminated=False):
     checkpoint.has_terminated = hasTerminated
     alg = checkpoint
     print('Last checkpoint at generation %i' % len(alg.callback.data['var']))
-    
-    # Update any changes made to the algorithms between runs 
+
+    # Update any changes made to the algorithms between runs
     from pymooCFD.setupCFD import n_ind
     alg.pop_size = n_ind
+    # alg.set
     return alg
 
 
@@ -75,7 +95,7 @@ def loadTxt(fileX, fileF, fileG=None):
         # G = np.loadtxt(f'{dataDir}/{fileG}')
         G = np.loadtxt(fileG)
     else:
-        G = None 
+        G = None
 
     from pymoo.model.evaluator import Evaluator
     from pymoo.model.population import Population
@@ -83,17 +103,17 @@ def loadTxt(fileX, fileF, fileG=None):
     # now the population object with all its attributes is created (CV, feasible, ...)
     pop = Population.new("X", X)
     pop = Evaluator().eval(StaticProblem(problem, F=F, G=G), pop)
-    
+
     from pymooCFD.setupOpt import n_ind
     # from pymoo.algorithms.so_genetic_algorithm import GA
     # # the algorithm is now called with the population - biased initialization
     # algorithm = GA(pop_size=n_ind, sampling=pop)
     from pymoo.algorithms.nsga2 import NSGA2
     algorithm = NSGA2(pop_size=n_ind, sampling=pop)
-    
+
     return algorithm
-    
-    
+
+
 
 
 # def archive(dirName, archName = 'archive.tar.gz'):
